backend/pyproject.toml
# pyproject.toml
[tool.poetry]
name = "prompt-platform-backend"
version = "0.1.0"
description = "Backend for Prompt Compiler (FastAPI + PostgreSQL)"
authors = ["Veer <you@example.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.100.0"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
SQLAlchemy = "^2.0.30"
asyncpg = "^0.27.0"
alembic = "^1.11.1"
pydantic = "^1.10.12"
python-jose = "^3.3.0"
passlib = "^1.7.4"
authlib = "^1.2.0"
python-dotenv = "^1.0.0"
httpx = "^0.24.1"
aioredis = "^2.0.1"
typing-extensions = "^4.8.0"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
black = "^23.9.0"
ruff = "^0.12.0"

[tool.poetry.scripts]
start = "uvicorn app.main:app --host 0.0.0.0 --port 8000 --reload"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
